stagingfinanzgurutosf/requirements.txt
asn1crypto==1.5.1
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.0
filelock==3.9.0
flake8==6.0.0
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
numpy==1.24.1
openpyxl==3.0.10
oscrypto==1.3.0
packaging==23.0
pandas==1.5.2
pluggy==1.0.0
pycodestyle==2.10.0
pycparser==2.21
pycryptodomex==3.16.0
pyflakes==3.0.1
PyJWT==2.6.0
pyOpenSSL==22.1.0
pytest==7.2.0
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.2
six==1.16.0
snowflake-connector-python==2.9.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.14
